Refract - move impl from SerializeAST.cc into RefractAST.[h|cc]
